+++ plume/src/test/kotlin/io/github/plume/oss/extractor/UpdateGraphTest.kt
@@ -47,12 +47,12 @@ class UpdateGraphTest {
     fun testGraphUpdate() {
         // Initial projection
         listOf(testFile1, testFile2).forEach { extractor.load(it) }
-        extractor.project().postProject()
+        extractor.project()
         val g1 = driver.getWholeGraph()
         // Update file and do an update projection
         testFile2 = rewriteFileContents(testFile2, testFile2Update)
         listOf(testFile1, testFile2).forEach { extractor.load(it) }
-        extractor.project().postProject()
+        extractor.project()
         val g2 = driver.getWholeGraph()
         val literalsG1 = g1.nodes().asSequence().filterIsInstance<Literal>().toList()
         val literalsG2 = g2.nodes().asSequence().filterIsInstance<Literal>().toList()